% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hcov.R
\name{Hcov}
\alias{Hcov}
\title{Kernel bandwidth selection method based on bivariate density
contours coverage}
\usage{
Hcov(Y, shape = seq(1, 10, 0.5), k = 20, tau = 0.9, display_plot = TRUE)
}
\arguments{
\item{Y}{A matrix containing bivariate data values.}

\item{shape}{A sequence of values which controls plug in estimator increasing.}

\item{k}{A number indicating k fold cross validations to be performed.}

\item{tau}{The desired region coverage}

\item{display_plot}{A logical indicating if a plot must be displaying, during
the function estimation process, summarizing the results.}
}
\value{
This function return a diagonal kernel bandwidth matrix.
}
\description{
This function implements a method for estimating bivariate kernel bandwidth
based on data covarage. The method starts with the plug in estimate (which
usually overfits the data), and then increase this bandwidth value until the
desired coverage is obtained. Region coverage is evaluated in an out sample
design, using a k fold cross validation scheme.
}
