% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bivRegr.R
\name{plot.bivRegr}
\alias{plot.bivRegr}
\title{Plot method for bivRegr fit}
\usage{
\method{plot}{bivRegr}(x, eq = 1, newdata = NULL, ...)
}
\arguments{
\item{x}{A bivRegr fit.}

\item{eq}{A number indicating the model effects to be depicted; 1 = first response mean,
2 = second response mean, 3 = first response variance, 4 = second response variance, and
5 = correlation model.}

\item{newdata}{An optional data frame containing covariates values.}

\item{...}{Additional plot arguments.}
}
\value{
This function returns a plot for bivRegr mean, variance and correlation models.
}
\description{
This function takes an bivRegr object and plots the estimated effects for the
conditional response means, variances or correlation. summary_boot function
must be applied by the user in order to get the estimated effects confidence
intervals.
}
