% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.refcurv.R
\name{plot.refcurve}
\alias{plot.refcurve}
\title{Plot univariate conditional quantile models curves (i.e. reference curves)}
\usage{
\method{plot}{refcurve}(
  x,
  newdata = data.frame(x = seq(0, 1, 0.01)),
  tau = seq(0.1, 0.9, 0.2),
  ...
)
}
\arguments{
\item{x}{A refcurv object.}

\item{newdata}{A data frame defining a sequence of the predictor variables values.}

\item{tau}{A number or vector defining desired quantile.}

\item{...}{Additional plot options.}
}
\value{
This function returns a plot of the refcurve model.
}
\description{
This function depict the univariate conditional quantile model based
on the non parametric location scale model fitted with the refcurv function.
}
