% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.summary_boot.R
\name{plot.summary_boot}
\alias{plot.summary_boot}
\title{Default summary_boot plotting}
\usage{
\method{plot}{summary_boot}(x, eq = 1, select = NULL, ...)
}
\arguments{
\item{x}{A summary_boot object.}

\item{eq}{A number indicating the model effects to be depicted; 1 = first response mean,
2 =  second response mean, 3 = first response variance, 4 = second response variance, and
5 = correlation model.}

\item{select}{An optional parameter to represent an specific effect for each equation.}

\item{...}{Additional plot parameters, not yet implemented.}
}
\value{
This function returns a ggplot2 plot for the estimated effects along with
bootstrap 95\% confidence intervals.
}
\description{
This function takes the bivRegr bootstrap replicates obtained with summary_boot
function, and plots the parametric, and smooth effects for each model.
}
