% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bivRegr.R
\name{predict.bivRegr}
\alias{predict.bivRegr}
\title{Predict method for bivRegr}
\usage{
\method{predict}{bivRegr}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A bivRegr fit.}

\item{newdata}{A data frame defining the covariate values for prediction. Default
is NULL and the prediction will be done in the original data.}

\item{...}{Additional predict options.}
}
\value{
This function returns prediction of bivRegr mean, variance and correlation models.
}
\description{
Obtains predictions for a bivRegr object, that is the bivRegr means,
variances, and correlation for given covariate values.
}
