% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bivRegion.R
\name{summary.bivRegion}
\alias{summary.bivRegion}
\title{bivRegion summary method}
\usage{
\method{summary}{bivRegion}(object, tau = 0.95, ...)
}
\arguments{
\item{object}{A bivRegion object.}

\item{tau}{The data coverage proportion previously obtained by the bivRegion function.}

\item{...}{Additional summary options.}
}
\value{
This function indicates the region apparent coverage, and which
patients are located outside the estimated reference region.
}
\description{
This function takes an bivRegion object and indicates which observations are
located outside the estimated reference region and why.
}
