% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_boot.R
\name{summary_boot}
\alias{summary_boot}
\title{bivRegr summary function}
\usage{
summary_boot(object, B = 100, parallel = FALSE, cores = NULL)
}
\arguments{
\item{object}{A bivRegr fit.}

\item{B}{A number indicating the bootstrap iterations.}

\item{parallel}{A logical indicating if bootstrap parallelization must be
applied.}

\item{cores}{If parallel = TRUE, a number indicating computer cores
to be used during paralellization. If NULL the function use all cores available
but one.}
}
\value{
This function returns the bootstrap replicates of the bivRegr
sub models. Results might be checked applying plot.summary_boot().
}
\description{
This function perform a bootstrap procedure in order to obtain
confidence intervals for the bivRegr estimated effects. The function allow
to paralelize the bootstrap resampling scheme using doParalell, and
foreach libraries.
}
