% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivRegion.R
\name{trivRegion}
\alias{trivRegion}
\title{Trivariate reference region estimation}
\usage{
trivRegion(fit, tau = 0.9)
}
\arguments{
\item{fit}{A trivRegr object.}

\item{tau}{A number defining the desired coverage of the trivariate
reference region.}
}
\value{
This function return a region containing a given percentage of trivariate data points.
}
\description{
This functions estimate a probabilistic/reference region for trivariate data.
It is based on a non parametric kernel density estimation. It can only be applied to
a trivRegr object, and for one single tau.
}
\references{
Duong, T. (2019) ks: Kernel Smoothing. R package version 1.11.6. https://CRAN.R--project.org/package=ks.
}
